# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KRS Service API",
        "description": "API documentation for the Course & KRS Registration Service",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/v1/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Display a listing of courses and their remaining quota",
                "operationId": "371d9a18f115629527a67773ba85770a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Courses retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "IF-101"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Pemrograman Dasar"
                                                    },
                                                    "credits": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "quota": {
                                                        "type": "integer",
                                                        "example": 30
                                                    },
                                                    "remaining_quota": {
                                                        "type": "integer",
                                                        "example": 30
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-06-02T07:50:50Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-06-02T07:50:50Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "count": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized access. X-IAE-KEY header is missing or invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "auth": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Invalid API Key."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Header key for API authentication",
                "name": "X-IAE-KEY",
                "in": "header"
            },
            "BearerAuth": {
                "type": "http",
                "description": "Enter Bearer JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}